[indirect-executor] re-organise crate
